ahh so if we reference a cell then remove the reference it does not consider it an empty cell
